storage/requirements.txt
google-cloud-storage
google-cloud-pubsub
requests
flask
flask-cors